refactor: Add better logging, and make code more DRY
