+++ vortexasdk/endpoints/cargo_movements_result.py
@@ -10,7 +10,7 @@ from vortexasdk.api.entity_flattening import (
     convert_cargo_movement_to_flat_dict,
 )
 from vortexasdk.api.search_result import Result
-from vortexasdk.create_dataframe import create_dataframe
+from vortexasdk.result_conversions import create_dataframe, create_list
 from vortexasdk.logger import get_logger
 
 logger = get_logger(__name__)
@@ -26,13 +26,8 @@ class CargoMovementsResult(Result):
 
     def to_list(self) -> List[CargoMovement]:
         """Represent cargo movements as a list of `CargoMovementEntity`s."""
-        list_of_dicts = super().to_list()
-
-        with Pool(os.cpu_count()) as pool:
-            logger.debug(
-                f"Converting dictionary to CargoMovements using {os.cpu_count()} processes"
-            )
-            return list(pool.map(CargoMovement.from_dict, list_of_dicts))
+        # noinspection PyTypeChecker
+        return create_list(super().to_list(), CargoMovement)
 
     def to_df(self, columns=None) -> pd.DataFrame:
         """
