+++ vortexasdk/endpoints/timeseries_result.py
@@ -1,5 +1,3 @@
-import os
-from multiprocessing.pool import Pool
 from typing import List
 
 import pandas as pd
@@ -7,7 +5,7 @@ import pandas as pd
 from vortexasdk.api.search_result import Result
 from vortexasdk.api.timeseries_item import TimeSeriesItem
 from vortexasdk.logger import get_logger
-from vortexasdk.create_dataframe import create_dataframe
+from vortexasdk.result_conversions import create_dataframe, create_list
 
 logger = get_logger(__name__)
 
@@ -17,12 +15,8 @@ class TimeSeriesResult(Result):
 
     def to_list(self) -> List[TimeSeriesItem]:
         """Represents time series as a list."""
-        list_of_dicts = super().to_list()
-        with Pool(os.cpu_count()) as pool:
-            logger.debug(
-                f"Converting dictionary to TimeSeries using {os.cpu_count()} processes"
-            )
-            return list(pool.map(TimeSeriesItem.from_dict, list_of_dicts))
+        # noinspection PyTypeChecker
+        return create_list(super().to_list(), TimeSeriesItem)
 
     def to_df(self, columns=None) -> pd.DataFrame:
         """Represents the timeseries as a dataframe.
