+++ vortexasdk/endpoints/vessels_result.py
@@ -1,5 +1,3 @@
-import os
-from multiprocessing import Pool
 from typing import List
 
 import pandas as pd
@@ -7,7 +5,7 @@ import pandas as pd
 from vortexasdk.logger import get_logger
 from vortexasdk.api import Vessel
 from vortexasdk.api.search_result import Result
-from vortexasdk.create_dataframe import create_dataframe
+from vortexasdk.result_conversions import create_dataframe, create_list
 
 logger = get_logger(__name__)
 
@@ -17,13 +15,8 @@ class VesselsResult(Result):
 
     def to_list(self) -> List[Vessel]:
         """Represent vessels as a list."""
-        list_of_dicts = super().to_list()
-
-        with Pool(os.cpu_count()) as pool:
-            logger.debug(
-                f"Converting dictionary to Vessels using {os.cpu_count()} processes"
-            )
-            return list(pool.map(Vessel.from_dict, list_of_dicts))
+        # noinspection PyTypeChecker
+        return create_list(super().to_list(), Vessel)
 
     def to_df(self, columns=None) -> pd.DataFrame:
         """
